+++ lib/src/romanize_base.dart
@@ -7,6 +7,7 @@ enum RomanizerSystem {
   arabic,
   chinese,
   cyrillic,
+  devanagari,
   greek,
   hebrew,
   japanese,